Create `esproposal.decorators=warn|ignore` config option
Summary: This adds `esproposal.decorators` for users to turn on parsing of decorators with Flow. The benefit to this is that people can still opt in to using Flow when they also use decorators -- in spite of the fact that Flow doesn't understand decorators. To do this, users must explicitly specify either `ignore` or `warn` as the value of the option. The former skips over decorators and the latter gives a warning (presumably to be suppressed explicitly in code with something like `/*FlowIgnore*/`).

Reviewed By: @gabelevi

Differential Revision: D2412388