Point location of function declaration types at the function's name
Summary: Currently, when you extract the type information for a function declaration, the type's loc spans the entire declaration. In terms of UIs that highlight or have tooltips for types within code (editors, etc), this means you see the function's type is shown when hovering over unrelated stuff in the function's body.

This adjusts the location of the type in the type table to only span the function's name -- which is the most likely place one might hover in an editor to see the type description for a function declaration.

Reviewed By: @gabelevi

Differential Revision: D2418043