Refactor Env_js.bind_entry
The previous logic was pretty twisted--a side effect of my tacking on
some extra logic hastily.

The key thing here is that we need to actually check for existing
entries in a lex scope, even if we're trying to add a var-scoped
binding, because any shadowing in a lex scope is bad.

Also, it turns out that I didn't handle types shadowing lexicals, so
bully for that.