Abnormals should pop lex scope in catch clause
This resolves an issue which I misdiagnosed, causing me to add
unnecessary cases to Env_js.merge_env, which could mask other real
errors.

What's happening here: We enter a catch clause, which pushes a new lex
scope. Before we can pop the lex scope, we hit an abnormal and jump
away. Then, we merge the env that has a dangling lex scope on top.

In general, we run this risk whenever we push a lex scope and then
process statements that might raise. I have convinced myself that no
other code that calls merge_env will hit this specific case, but a
generalized approach would be nice, too.