Fix interaction between optional params and rest params
Summary: When we ran out of function arguments and the only parameter left was
a rest param, we would do the right thing. However, if we had something other
than a single rest param, we would flow `undefined` to each of the remaining
parameters. This meant that we were sometimes flowing `undefined` to a rest
parameter and that was wrong.

Fixes #789

Reviewed By: @jeffmo

Differential Revision: D2425046