Switch to disjoint union IteratorResult<Yield, Return>
Summary: When you iterate through an `Iterator<T>` by calling `next()`, you
get back `IteratorResult<T|void>`, which means when you access the `value`
property of that `IteratorResult` it can either be `T` or `void`.

Our interfaces use structural subtyping, which is cool and all except that it
has to unify the properties. That means an `IteratorResult<string>` is NOT a
subtype of `IteratorResult<string|void>`. Long story short, if you ever need to
deal with the `IteratorResult` of an `Iterator<string>`, you'll need to write
`IteratorResult<string|void>`

Switching to a disjoint union magically makes everything better, especially with some of the awesome boolean literal logic that @mroch has been adding.

Reviewed By: @jeffmo

Differential Revision: D2426119