Support `return` and `throw` generator methods
Per the spec[1], generator objects have two methods which I neglected to
implement on the first go 'round.

`return` is somewhat special, because it means the return type can be
widened outside the generator function declaration itself. The argument
passed to `return` becomes the completion value of the generator itself,
and return value is the iterator result with `done:true` and the
provided value.

`throw` is more straightforward, but led me to a (perhaps controversial)
change: I think the type of the catch parameter should be `mixed`,
instead of an open tvar. In a perfect world, Flow would know exactly
what type was thrown, but barring that, we should be pessimistic and
require the user to do a dynamic type test (`invariant()` is useful for
this).

This change is unrelated to generators, but I think the throw API
invites people to pass values into a generator, and I don't want Flow to
ignore potential runtime errors caused by that use.

1. http://www.ecma-international.org/ecma-262/6.0/#sec-generator-objects