multiple inheritance
Summary: Support interfaces with multiple extends.
Also support mixins for declare classes.

Together, these features boost Flow's ability to describe libraries via declarations.

An interface should be used when there is no runtime value by that name (the thing being expressed is purely a type). Multiple inheritance is available via 'extends'.

A declare class should be used when a runtime value is being described, and we also need to describe a type by the same name. Multiple inheritance is available via a new clause declaring 'mixins': `declare class Foo extends Bar mixins Qux`.

'implements' is not implemented yet, since it is redundant in a world where interfaces are matched structurally. But we may implement it in the future, as a code structuring mechanism (which may also speed up type checking).

Reviewed By: @jeffmo

Differential Revision: D2292036