Try to infer private/munged members on exported classes
Summary: The notion of requiring an annotation on an export is in place to improve documentation of module interfaces and to reduce the depth of the graph to be solved during merge time. However, when the class-munging/private properties feature is turned on, these methods are neither public APIs nor are they likely to be used or referenced outside the module -- so auto-inferring them seems like the right thing to do.

Reviewed By: @mroch

Differential Revision: D2448146