dep calc fix for incr recheck bug
Summary: (In which I was being too clever.)

We've gone back and forth on the use of memoization for calculating dependents
for remerging. It caused problems before, @bhosmer took it out, and I put it
back with a slight twist. Taking it out again because it occasionally causes
intermediate dependents to be missed (e.g., when they are part of a cycle),
which causes merging to crash.

Why does memoization not work? Because as the computation is set up, results can
go from false to true: we're searching for the existence of paths between nodes
in a graph via DFS, and at some point a path shows up, even though previous
recursive calls that explored a part of the graph may not have discovered a
path.

Reviewed By: @bhosmer

Differential Revision: D2450603