Fix type widening for switch statements without break statements
Summary: In response to #786, in switch statements without any break statements (that is, every branch falls through) where there is a default case, the type refinements from the different case branches are not preserved.

The cause for this is that after gathering type information from the branches, we wipe the changeset and have a special case for switch statements without default statements that has the side effect of adding in any of the type information we were going to wipe.  Additionally, if there was a break statement at all in the code, we havoc the variables.  This happens to leave out the case where we have a default and every case falls through.
Closes https://github.com/facebook/flow/pull/792

Reviewed By: @avikchaudhuri

Differential Revision: D2453610

Pulled By: @bhosmer