Allow legal const / uninitialized destructuring uses
There are a couple of separate issues here:

1. Destructuring assignment used Env_js.set_var, which is intended for
modification and prevents initializing const-bound destructured
patterns. I fixed this by using the same `destructuring` helper, but
using the appropriate `Env_js.init_*` function instead of `set_var`.

2. The parser rejects destructuring patterns without initialization in
positions where it is legal. Specifically in for-of and for-in loops,
where the initialization is implied. I fixed this by adding an
`implicit_init` state to the parser.

For parsing, it's a bit complicated because we don't know if init is
implied until we reach the `T_IN` or `T_OF` token. I employed
backtracking here, so we first parse as if init were implied, but if it
turns out we are not in a for-in/for-of, we start over without implied
init, because it's a plain for loop. Is there a more efficient method?