Fix incremental make on OSX when lib files change
Summary: @​danvk noticed that when you changed the lib files and than ran `make` (without a `make clean`) that the resulting binary would still use the old libraries. I looked into it and found an issue with incremental make on OSX.

Long story short, we want ocamlbuild to re-link the binary when the lib files change, since on OSX we rely on ocamlbuild to embed the lib files in the binary (`objcopy` is not available). However ocamlbuild doesn't understand that it needs to relink when non-ocaml dependencies change.
Closes https://github.com/facebook/flow/pull/840

Reviewed By: @​svcscm

Differential Revision: D2464340

Pulled By: @gabelevi