improve error positions
Summary: repositions types as they flow through certain operations, in order to improve the locality of errors.

adds a `ReposLowerT of t` use type which waits for the lower bound to become concrete, then repositions it to the location of the type being carried around in the `ReposLowerT`. this type is used instead of most of the places that we were using `repos_t_from_reason` before, so that the repositioning moves a concrete type instead of potentially moving a tvar, which is useless since the tvar's position does not impact the types flowing through it. it's used for variable declaration, identifier lookups (`var_ref`), function calls (flowing params to args) and method calls (same... but we weren't doing this before, not even with `repos_t_from_reason`!).

the effect is that errors are significantly more local... each side of the equal sign, for example, instead of wherever the types that flow into each side of an assignment were originally defined.

Reviewed By: @​@bhosmer

Differential Revision: D2448168