tweak position of initialState
Summary: React calls `getInitialState` internally and we seem to pretend that it creates an `initialState` variable. this diff tweaks the location of that call so that error messages involving that function appears to have happened at the return position of the function, rather than the entire body. we do basically the same thing for `getDefaultProps` immediately above `getInitialState`.

Reviewed By: @​@​avikchaudhuri

Differential Revision: D2463136