change filenames from string to variant
Summary: creates a `filename` variant consisting of `LibFile`, `SourceFile`, or `Builtins`. using this in place of strings lets us tag whether a filename is, for example, a library or the global pseudo-file, so we don't have to check against the list of lib files (notably, the client doesn't know the list of lib files). Now that the client can tell what is a lib file, we should be able to make `--strip-root` a client option. We can also include the type in JSON instead of asking people to look for a magical "[LIB]" prefix.

Reviewed By: @bhosmer

Differential Revision: D2479174