track error kind, only add banner in cli
Summary: we add a `Library (parse|type) error:` banner to errors that happen in lib files to draw attention to them, but this breaks the pattern that the first line is the type of the thing that failed (op banners have the same problem but will deal with those separately). this diff, along with the change to `Loc.filename`, adds enough info to errors to add this banner after the fact, only in the CLI. a JSON client could do the same if they wanted to, but I don't think it necessarily makes sense there, when errors are being shown inline.

Reviewed By: @bhosmer

Differential Revision: D2481186