add --strip-root to status command
Summary: Now that the client can tell which paths are lib files is (via `Loc.filename`), we can strip paths client-side, which means we can add `--strip-root` to `status`!

If `strip_root=true` is in the `.flowconfig`, roots are always stripped (`--strip-root` has no effect).

also, traces still read `modes.strip_root`, which is only set on the server. so when using `traces=X` in `.flowconfig`, traces may be misaligned (extra padding) unless `strip_root=true` is set in the `.flowconfig`. the CLI flag will still strip the paths, but there may be extra spaces because the padding was based on the unstripped paths.

however, `flow check --strip-root --traces=X` WILL work, because that fires up a new "server" every time.

Reviewed By: @bhosmer

Differential Revision: D2481581