autocomplete despite parse fail
Summary: Autocomplete would fail for partial files that (with dummy token) didn't parse.

This diff relaxes parsing when typechecking contents. The rationale is that
since contents are not going to affect future typechecking, we can afford to
overlook parse failures and instead try producing meaningful results.

Autocomplete relies on typechecking contents, so benefits from the relaxation.
But so would other commands that rely on typechecking contents.

Reviewed By: @bhosmer

Differential Revision: D2470839