Lex object keys in normal mode
Summary: Fixes #829, where @sebmck couldn't use `implements` as the key in a
declared class type. This was happening because the key was lexed in
"type mode" which spat out a T_IDENTIFIER instead of a T_IMPLEMENTS, so
later on, in `Parser_flow.identifier_or_reserved_word`, the logic to
explicitly allow reserved words didn't fire, and we fell back to
`identifier`, which complains about the future reserved word.

After this change object keys in a type context are lexed in normal
mode, so the lexer will spit out the more specific token type, and
`identifier_or_reserved_word` can do its job.

This change does allow `@@​foo` in object literals and class
declarations, which is absolutely wrong. However, we should really be
using [Symbol.iterator] instead of @@​iterator, so adding a bunch of
error handling in the parser (tried it, was ugly) just didn't seem worth
it. Feel free to set me straight if I'm just being lazy... :)
Closes https://github.com/facebook/flow/pull/870

Reviewed By: @jeffmo

Differential Revision: D2486482

Pulled By: @gabelevi