Standardize exit codes
Summary: This diff attempts to create a list of the various exit codes and use
them in the right places.

Exit codes are part of our API and therefore can't be changed willynilly.
Luckily, our exit codes are mostly undocumented, and only a few are used by the
various tools that integrate with Flow. This gives me the freedom to change
with which code many paths exit.

Reviewed By: @bhosmer

Differential Revision: D2493921