array spread fix
Summary: Earlier, we used to treat spreads in arrays the same way as we treat spreads in
call args. This was flawed, because in call args they implicitly repeat (to
match a sequence of params) whereas in arrays they don't: all elements have the
same type.

Matching with repetition needs special handling, which we implement only in
parameter / call arg matching. This diff frees spreads in arrays from
repetition, so that no further special handling is needed for them.

Reviewed By: @bhosmer

Differential Revision: D2497212