All classes have a static "name" property
Summary: All classes implicitly have a static property `name`, which is the
name of the class. Anonymous classes have the empty string. Initially I built
this to use string literals, but this doesn't work well with inheritance. That
is if `class B extends A {}`, then it's a problem if `B.name` and `A.name` have
incompatible string literal types.

Functions have this too, but at the moment we don't enforce anything about the
statics of functions.

Reviewed By: @bhosmer

Differential Revision: D2497035