[flow parser] Fix 'value' property of T_TEMPLATE_PART tokens to include backtick, "${", and "}"
Summary: Previously the T_TEMPLATE_PART token passed up from the lexer emitted the "raw" (in esprima terminology) value of AST node.
The problem with this is that the data in "raw" doesn't include interlude chars like ```, `${`, and `}`.

This fixes the lexer to include that "really raw" stuff in the token in addition to what it included previously. This allows a token sink to receive a "value" field that matches the "range" that's specified for the token.

Reviewed By: @gabelevi

Differential Revision: D2499443