record top/any/bot in bounds for consistency
Summary: In the past, we found it useful for perf to omit top/any/bot types in
upper/lower and upper/lower bounds of tvars. Unfortunately, while this makes
sense for typechecking (in the sense that such constraints have no effect on
typechecking), it caused unintended effects: e.g., we couldn't distinguish tvars
that were never defined/used from those that were only defined/used with any.

It turns out that this hack doesn't boost perf any more (due to improvements
made elsewhere), so this diff simplifies the model for other (future) benefits
such as detecting tvars that are never defined/used (and others...like deciding
whether an inferred type is "polluted" by any, which could be useful for
implementing sound tranforms).

Reviewed By: @bhosmer

Differential Revision: D2501262