Allow reserved keywords in import/export when safe to do so
Summary: If we are doing a delegate export, the identifiers in the export
specifiers come from and go to object keys, so they follow the same
rules as object keys. If there is no "from" in the export statement, the
identifiers come from bindings, and thus follow those rules.

Import statements can rebind identifiers that come from object keys to
valid binding identifiers. If there is an "as" in the import binding
list item, the left-hand identifier can be a reserved keyword.

In order to allow the parse with the option of raising later, I modified
the signature of `identifier_or_reserved_keyword` to return the parsed
identifier along with the error that would have been raised. I factored
out the logic to determine this error from `error_unexpected`.

Fixes #828, #823
Closes https://github.com/facebook/flow/pull/896

Reviewed By: @mroch

Differential Revision: D2501816

Pulled By: @gabelevi