don't hang when merge throws
Summary: Earlier the merging phase would hang if a particular merge job crashed. This is
because the merge stream's join function would have no way to notify that
dependents of the files involved in the crashed merge job should be removed from
the queue.

This diff fixes that situation. Errors are collected for only successful merge
jobs, but join works for even for failed merge jobs.

Reviewed By: @gabelevi

Differential Revision: D2503843