Use smaller max depth for printing raw type
Summary: The JSON serialization was causing nontermination (or at least taking an absurd
amount of time). This seems to fix it.

I found this out by printing to stdout before and after the serialization and
observing that after messing around in Nuclide, it would eventually print the
before message but not the after message.

Reviewed By: @bhosmer

Differential Revision: D2505263