[hack] Mini saved state
Summary: This is the first bit of saved state v2. We download the dep table in parallel
during parsing. We don't actively skip any typechecking steps, so any speedup
would be caused by writes to the dependency table being no-ops (because the
entry already exists).

Running this would give us a sense of how long it takes to download smaller
packages -- I'm particularly interested to know if we can usually complete the
download + load before parsing completes.

This is gated on Chef.

Reviewed By: @dlreeves

Differential Revision: D2502466