Fix new error format for strip-root
Summary: The new error format would print blank lines if it couldn't find the
file in question. If you ran flow with --strip-root, the errors would use
relative paths. If you were also not in the root directory, then the relative
paths wouldn't lead to a file.

This diff makes a few improvements:

* If the paths are relative, then we look them up relative to the root
* Currently, strip-root turns `/tmp/flow/flowlib_10002a87/core.js` to `[LIB] core.js` and `lib/foo.js` to `[LIB] foo.js`. This diff changes the latter to `[LIB] lib/foo`. So lib files within the root get a relative path rather than just a basename.
* If we're dealing with a lib file and if it starts with `[LIB] ` then we strip that out.
* If for some reason the file doesn't exist, then we degrade gracefully to just showing the error message and the file name. So flowlib files show up as `[LIB] core.js`
* If the file doesn't exist but we're using stdin instead, then no need to degrade

Ideally we'd be able to always extract the flowlib to a deterministic relative path when running the tests, so we could always provide a relative path. I looked into this a little more, but it would involve some options and is probably out of the scope of this diff

Reviewed By: @bhosmer

Differential Revision: D2513946

fb-gh-sync-id: 55b473a578f561eb8e74069c0f16b03030d8c34a