add parser option to disable parsing types
Summary: adds a flag to disable type parsing, which makes it an error if you use `type`, `interface`, `implements`, annotations or type variables. this could be useful, for example, if using the Flow parser independently from Flow, or when Flow is parsing non-`@​flow` files.

Today, a non-`@​flow` file can have types, with no syntax errors, even though the types are completely ignored. I want to use this mode to disable types in those files, but allow types in `@​noflow` files (or maybe rename that to something else, like `@​flow disabled`). Otherwise, making some things opt-out and other things opt-in based on the flowconfig instead of the docblock will give you no indication of whether the types you're writing are actually doing anything. So, if you're in an opt-in file and don't opt in, I want the annotations to error.

Reviewed By: @jeffmo

Differential Revision: D2511130

fb-gh-sync-id: 2eda4e76b73c4912ce6575e78666d5109be42090