Add support for a `<PROJECT_ROOT>` template token to the module.name_mapper config option
Summary: This will allow specification of a `module.name_mapper` config template that is relative to the project root. This is useful in some circumstances where name_mappers are used to customize aspects of the module resolution algorithm.

For example, in a hypothetical project, module name resolution might vary from Node's resolution algorithm such that the lookup on the filesystem starts in a `src` directory at the root of the project. To accomplish this, one could specify `module.name_mapper='^\(.*\)$' -> '<PROJECT_ROOT>/src/\1'` to canonicalize an import of `"foo/bar/baz"` to `"/path/to/root/of/project/src/foo/bar/baz"`.

Reviewed By: @gabelevi

Differential Revision: D2518465

fb-gh-sync-id: df4c0729e0f68207eb013c00004377aa3d3ffa23