isolate async tests from unrelated bug
Summary: it would appear that `Env_js.get_var` returns the last polymorphic type by a given name, rather than the one you want. In this test file, note how there is a `Promise<void>` and a `Promise<number>` in the global scope. It would seem that when `await` looks up `Promise` in order to fake a call to `resolve` on it, we get back the last one with its type params, rather than the one we expected (actually in `test1` we probably expect one without any types applied, I guess?)

this diff isolates each test from the others to avoid this problem. note how the first error changes.

Reviewed By: @gabelevi

Differential Revision: D2502285

fb-gh-sync-id: b6f3d0a20f75e7514c44dbf4f35387ae70e3beec