add accessors/mutators to Context
Summary: makes `Context.t` opaque, and adds public accessors and mutators.

the reason here is twofold:

firstly, this DRYs up the `cx.foo <- IMap.add ... cx.foo` boilerplate and some `cx.Context.foo` ugliness (not sure `(Context.foo cx)` is that much better because of the parens, but it's more readable...
secondly, I want to create a config record containing the fields you need to pass to create a context, instead of a ton of params to `new_context`, but ocaml won't let me use the same field names, so I'm going to store the whole config in the context directly, and trivially update the accessors.

Reviewed By: @bhosmer

Differential Revision: D2525538

fb-gh-sync-id: 70524bcff915d3972a41e142591c0685c0a0513c