stricter null
Summary: Earlier we modeled the type of `null` as NullT | X for some X. The hope was that
this would allow objects with null properties to flow to object types where
those properties had nullable types. But this is already possible due to our
special subtyping rules for literals. Moreover, the additional X would cause
missing annotation errors when `null` was part of exports.

This diff simplifies the model by typing `null` as NullT. We lose some inference
due to this, but the inference was shallow (both literally and figuratively)
anyway, and new errors are easily fixed by type-casting `null` to the desired
nullable type.

This also paves the way for a `null` singleton type (= NullT).

Reviewed By: @bhosmer

Differential Revision: D2500266

fb-gh-sync-id: 70ececb2f7438268d5455f750a336b53e5d725c5