anyfun anyobj typeof refi
Before this change, the two added test cases resulted in a type error,
because the conditional did not result in a refinement. This is because
`Function` and `Object` annotations are represented as `AnyFunT` and
`AnyObjT` internally, respectively. These types were erroneously
excluded from the pattern match `is_function` and `is_object`,
respectively.

Fixes #936