Refactor option parsing in FlowConfig
Summary: In D2462248 we had to one-off hack the option parser logics to provide a default value for `module.file_exts` when the user has specified *no* custom file extensions, but initialize the list of file extensions to an empty list if the user provided any custom file extensions.

In order to include this "initializer value" in the old `option_entry` record, that record needed to be parameterized on the type of the option value. However, if we do this we end up building a list of polymorphic elements -- which ocaml can't deal with in any reasonably ergonomic way I could think of.

So I went in and refactored option definitions into a chain of functions rather than storing into a map. This lets each option definition define it's own string -> value parser, an optional initializer value (for things like `module.file_exts`), and ultimately anything else that's parametric over the type of the resulting config option.

In the process I also cleaned a few crufty things up and moved option parsing into a more isolated submodule. I did consider moving the `Opts` module into it's own file, but it got a little bit hairy with refs -- so ultimately I landed on `FlowConfig.Opts` instead.

Reviewed By: @gabelevi

Differential Revision: D2527803

fb-gh-sync-id: 82d1941e2828faabf45346b0e660d5384ce691a6