[hack] Improve Typing of TypeStructure<T>
Summary: Currently we add some specialization to the `TypeStructure` type when we
localize decl ty -> locl ty. However this means many things won't work as you
expect with the type, in particular around instantiating generics.

The solution I've come up with is to remove the specialization from
Typing_tdefs expansion and instead perform it whenever we index into a
`TypeStructure`.

Reviewed By: @dabek

Differential Revision: D2534747

fb-gh-sync-id: ea4eed2c38439d521ee087823134aa43854b8035