anyfun anyobj typeof refi
Summary: Before this change, the two added test cases resulted in a type error,
because the conditional did not result in a refinement. This is because
`Function` and `Object` annotations are represented as `AnyFunT` and
`AnyObjT` internally, respectively. These types were erroneously
excluded from the pattern match `is_function` and `is_object`,
respectively.

Fixes #936
Closes https://github.com/facebook/flow/pull/941

Reviewed By: @​svcscm

Differential Revision: D2545469

Pulled By: @gabelevi

fb-gh-sync-id: 192808c750ffbe0a245e20efb0fcc010cd117984