[hack] Acquire lock as early as possible
Summary: Previously, we would acquire the lock after doing some work, including
starting up watchman. In theory this should be okay because watchman has a lock
to prevent multiple concurrent servers, but it wasn't working as well as it
should. Let's move our lock to an earlier stage in our initialization as a
workaround.

Downside: You can do no longer do `hh_server --check` on a directory that has a
hh_server daemon monitoring it. I don't think it's a big deal, though.

Reviewed By: @​achow

Differential Revision: D2543288

fb-gh-sync-id: 505580fb42037742cf7bd6a17d8b19a58ebe92c3