Add version command
Summary: `flow --version` was a bit of a hack. It should really be its own
command. As it's own command we can add options, like `--json`. And in the json
output we can put more useful information, like where the binary is actually
located and even the build id.

We need to leave `flow --version` in for legacy support for a little while, so
for now we can just mark it as deprecated.

Reviewed By: jeffmo

Differential Revision: D2550602

fb-gh-sync-id: 361ce6c852ce7c837e30443bf94434575e5fba0c