enforce tdz on writes
Summary: Previously we were enforcing TDZ on reads to let/const, but not writes,
allowing errors like the following:

  function f(b) {
    x = 10;
    let x = 0
  }

This diff adds the same check to writes as we do for reads.
Note that we remain overly lenient on forward references
within closures.

Added test in tests/binding/scopes.js

Reviewed By: gabelevi

Differential Revision: D2554141

fb-gh-sync-id: 25e86f2f9d88220f1398c3ea48647dba9b28b623